login and sign up done

+++ src/routes/post/index.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const { PostsModel } = require("./schema");
+const ProfileModel = require("../profiles/schema");
 const router = express.Router();
 const multer = require("multer");
 const upload = multer({});
@@ -36,7 +37,8 @@ router.get("/:id", async (req, res, next) => {
 });
 router.post("/", async (req, res, next) => {
   try {
-    const newPost = new PostsModel(req.body);
+    const user = await ProfileModel.findOne({ username: req.body.username });
+    const newPost = new PostsModel({ ...req.body, user: user._id });
     const { _id } = await newPost.save();
     //{ runValidators: true }
 
@@ -63,7 +65,9 @@ router.put("/:id", async (req, res, next) => {
 
 router.delete("/:id", async (req, res, next) => {
   try {
-    await fs.unlink(join(imgFolderPath, `${req.params.id}.png`));
+    if (fs.existsSync(join(imgFolderPath, `${req.params.id}.png`))) {
+      await fs.unlink(join(imgFolderPath, `${req.params.id}.png`));
+    }
     const post = await PostsModel.findByIdAndDelete(req.params.id);
     if (post) res.status(200).send("Deleted");
     else res.status(404).send("Not found");
@@ -79,6 +83,9 @@ router.post("/:id/upload", upload.array("avatar"), async (req, res, next) => {
       writeFile(join(imgFolderPath, `${req.params.id}.png`), file.buffer)
     );
     await Promise.all(arrayOfPromises);
+    const addImage = await PostsModel.findByIdAndUpdate(req.params.id, {
+      image: `http://localhost:${process.env.PORT}/img/posts/${req.params.id}.png`,
+    });
     res.status(200).send("uploaded");
   } catch (e) {
     next(e);
